requirements.txt
Django==4.2.2
grpcio-tools==1.56.0
grpcio==1.56.0
pycryptodome==3.18.0